backend/pyproject.toml
[project]
name = "career-mode-live"
version = "0.1.0"
description = "Multi-agent voice career coaching panel — Gemini Live Agent Challenge"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=1.0.0",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
